Refactor keyboard input modules
