+++ gambaterm/keys.py
@@ -1,0 +1,322 @@
+from __future__ import annotations
+
+from enum import Enum, auto
+
+
+class LatinKeys(Enum):
+    SPACE = " "
+    EXCLAM = "!"
+    QUOTEDBL = '"'
+    NUMBERSIGN = "#"
+    DOLLAR = "$"
+    PERCENT = "%"
+    AMPERSAND = "&"
+    QUOTERIGHT = "'"
+    PARENLEFT = "("
+    PARENRIGHT = ")"
+    ASTERISK = "*"
+    PLUS = "+"
+    COMMA = ","
+    MINUS = "-"
+    PERIOD = "."
+    SLASH = "/"
+    DIGIT_0 = "0"
+    DIGIT_1 = "1"
+    DIGIT_2 = "2"
+    DIGIT_3 = "3"
+    DIGIT_4 = "4"
+    DIGIT_5 = "5"
+    DIGIT_6 = "6"
+    DIGIT_7 = "7"
+    DIGIT_8 = "8"
+    DIGIT_9 = "9"
+    COLON = ":"
+    SEMICOLON = ";"
+    LESS = "<"
+    EQUAL = "="
+    GREATER = ">"
+    QUESTION = "?"
+    AT = "@"
+    CAPITAL_A = "A"
+    CAPITAL_B = "B"
+    CAPITAL_C = "C"
+    CAPITAL_D = "D"
+    CAPITAL_E = "E"
+    CAPITAL_F = "F"
+    CAPITAL_G = "G"
+    CAPITAL_H = "H"
+    CAPITAL_I = "I"
+    CAPITAL_J = "J"
+    CAPITAL_K = "K"
+    CAPITAL_L = "L"
+    CAPITAL_M = "M"
+    CAPITAL_N = "N"
+    CAPITAL_O = "O"
+    CAPITAL_P = "P"
+    CAPITAL_Q = "Q"
+    CAPITAL_R = "R"
+    CAPITAL_S = "S"
+    CAPITAL_T = "T"
+    CAPITAL_U = "U"
+    CAPITAL_V = "V"
+    CAPITAL_W = "W"
+    CAPITAL_X = "X"
+    CAPITAL_Y = "Y"
+    CAPITAL_Z = "Z"
+    BRACKETLEFT = "["
+    BACKSLASH = "\\"
+    BRACKETRIGHT = "]"
+    ASCIICIRCUM = "^"
+    UNDERSCORE = "_"
+    QUOTELEFT = "`"
+    A = "a"
+    B = "b"
+    C = "c"
+    D = "d"
+    E = "e"
+    F = "f"
+    G = "g"
+    H = "h"
+    I = "i"  # noqa: E741
+    J = "j"
+    K = "k"
+    L = "l"
+    M = "m"
+    N = "n"
+    O = "o"  # noqa: E741
+    P = "p"
+    Q = "q"
+    R = "r"
+    S = "s"
+    T = "t"
+    U = "u"
+    V = "v"
+    W = "w"
+    X = "x"
+    Y = "y"
+    Z = "z"
+    BRACELEFT = "{"
+    BAR = "|"
+    BRACERIGHT = "}"
+    ASCIITILDE = "~"
+    NOBREAKSPACE = "\xa0"
+    EXCLAMDOWN = "¡"
+    CENT = "¢"
+    STERLING = "£"
+    CURRENCY = "¤"
+    YEN = "¥"
+    BROKENBAR = "¦"
+    SECTION = "§"
+    DIAERESIS = "¨"
+    COPYRIGHT = "©"
+    ORDFEMININE = "ª"
+    GUILLEMOTLEFT = "«"
+    NOTSIGN = "¬"
+    HYPHEN = "\xad"
+    REGISTERED = "®"
+    MACRON = "¯"
+    DEGREE = "°"
+    PLUSMINUS = "±"
+    TWOSUPERIOR = "²"
+    THREESUPERIOR = "³"
+    ACUTE = "´"
+    MU = "µ"
+    PARAGRAPH = "¶"
+    PERIODCENTERED = "·"
+    CEDILLA = "¸"
+    ONESUPERIOR = "¹"
+    MASCULINE = "º"
+    GUILLEMOTRIGHT = "»"
+    ONEQUARTER = "¼"
+    ONEHALF = "½"
+    THREEQUARTERS = "¾"
+    QUESTIONDOWN = "¿"
+    CAPITAL_A_GRAVE = "À"
+    CAPITAL_A_ACUTE = "Á"
+    CAPITAL_A_CIRCUMFLEX = "Â"
+    CAPITAL_A_TILDE = "Ã"
+    CAPITAL_A_DIAERESIS = "Ä"
+    CAPITAL_A_RING = "Å"
+    CAPITAL_A_E = "Æ"
+    CAPITAL_C_CEDILLA = "Ç"
+    CAPITAL_E_GRAVE = "È"
+    CAPITAL_E_ACUTE = "É"
+    CAPITAL_E_CIRCUMFLEX = "Ê"
+    CAPITAL_E_DIAERESIS = "Ë"
+    CAPITAL_I_GRAVE = "Ì"
+    CAPITAL_I_ACUTE = "Í"
+    CAPITAL_I_CIRCUMFLEX = "Î"
+    CAPITAL_I_DIAERESIS = "Ï"
+    CAPITAL_E_TH = "Ð"
+    CAPITAL_N_TILDE = "Ñ"
+    CAPITAL_O_GRAVE = "Ò"
+    CAPITAL_O_ACUTE = "Ó"
+    CAPITAL_O_CIRCUMFLEX = "Ô"
+    CAPITAL_O_TILDE = "Õ"
+    CAPITAL_O_DIAERESIS = "Ö"
+    MULTIPLY = "×"
+    CAPITAL_O_OBLIQUE = "Ø"
+    CAPITAL_U_GRAVE = "Ù"
+    CAPITAL_U_ACUTE = "Ú"
+    CAPITAL_U_CIRCUMFLEX = "Û"
+    CAPITAL_U_DIAERESIS = "Ü"
+    CAPITAL_Y_ACUTE = "Ý"
+    CAPITAL_T_HORN = "Þ"
+    SSHARP = "ß"
+    A_GRAVE = "à"
+    A_ACUTE = "á"
+    A_CIRCUMFLEX = "â"
+    A_TILDE = "ã"
+    A_DIAERESIS = "ä"
+    A_RING = "å"
+    A_E = "æ"
+    C_CEDILLA = "ç"
+    E_GRAVE = "è"
+    E_ACUTE = "é"
+    E_CIRCUMFLEX = "ê"
+    E_DIAERESIS = "ë"
+    I_GRAVE = "ì"
+    I_ACUTE = "í"
+    I_CIRCUMFLEX = "î"
+    I_DIAERESIS = "ï"
+    E_TH = "ð"
+    N_TILDE = "ñ"
+    O_GRAVE = "ò"
+    O_ACUTE = "ó"
+    O_CIRCUMFLEX = "ô"
+    O_TILDE = "õ"
+    O_DIAERESIS = "ö"
+    DIVISION = "÷"
+    OSLASH = "ø"
+    U_GRAVE = "ù"
+    U_ACUTE = "ú"
+    U_CIRCUMFLEX = "û"
+    U_DIAERESIS = "ü"
+    Y_ACUTE = "ý"
+    T_HORN = "þ"
+    Y_DIAERESIS = "ÿ"
+
+    @classmethod
+    def from_latin(cls, code: int | str) -> LatinKeys | None:
+        try:
+            as_str = code if isinstance(code, str) else chr(code)
+            return cls(as_str)
+        except ValueError:
+            return None
+
+
+class FunctionalKeys(Enum):
+    ESCAPE = auto()
+    ENTER = auto()
+    TAB = auto()
+    BACKSPACE = auto()
+    INSERT = auto()
+    DELETE = auto()
+    LEFT = auto()
+    RIGHT = auto()
+    UP = auto()
+    DOWN = auto()
+    PAGE_UP = auto()
+    PAGE_DOWN = auto()
+    HOME = auto()
+    END = auto()
+    CAPS_LOCK = auto()
+    SCROLL_LOCK = auto()
+    NUM_LOCK = auto()
+    PRINT_SCREEN = auto()
+    PAUSE = auto()
+    MENU = auto()
+    F1 = auto()
+    F2 = auto()
+    F3 = auto()
+    F4 = auto()
+    F5 = auto()
+    F6 = auto()
+    F7 = auto()
+    F8 = auto()
+    F9 = auto()
+    F10 = auto()
+    F11 = auto()
+    F12 = auto()
+    F13 = auto()
+    F14 = auto()
+    F15 = auto()
+    F16 = auto()
+    F17 = auto()
+    F18 = auto()
+    F19 = auto()
+    F20 = auto()
+    F21 = auto()
+    F22 = auto()
+    F23 = auto()
+    F24 = auto()
+    F25 = auto()
+    F26 = auto()
+    F27 = auto()
+    F28 = auto()
+    F29 = auto()
+    F30 = auto()
+    F31 = auto()
+    F32 = auto()
+    F33 = auto()
+    F34 = auto()
+    F35 = auto()
+    KP_0 = auto()
+    KP_1 = auto()
+    KP_2 = auto()
+    KP_3 = auto()
+    KP_4 = auto()
+    KP_5 = auto()
+    KP_6 = auto()
+    KP_7 = auto()
+    KP_8 = auto()
+    KP_9 = auto()
+    KP_DECIMAL = auto()
+    KP_DIVIDE = auto()
+    KP_MULTIPLY = auto()
+    KP_SUBTRACT = auto()
+    KP_ADD = auto()
+    KP_ENTER = auto()
+    KP_EQUAL = auto()
+    KP_SEPARATOR = auto()
+    KP_LEFT = auto()
+    KP_RIGHT = auto()
+    KP_UP = auto()
+    KP_DOWN = auto()
+    KP_PAGE_UP = auto()
+    KP_PAGE_DOWN = auto()
+    KP_HOME = auto()
+    KP_END = auto()
+    KP_INSERT = auto()
+    KP_DELETE = auto()
+    KP_BEGIN = auto()
+    MEDIA_PLAY = auto()
+    MEDIA_PAUSE = auto()
+    MEDIA_PLAY_PAUSE = auto()
+    MEDIA_REVERSE = auto()
+    MEDIA_STOP = auto()
+    MEDIA_FAST_FORWARD = auto()
+    MEDIA_REWIND = auto()
+    MEDIA_TRACK_NEXT = auto()
+    MEDIA_TRACK_PREVIOUS = auto()
+    MEDIA_RECORD = auto()
+    LOWER_VOLUME = auto()
+    RAISE_VOLUME = auto()
+    MUTE_VOLUME = auto()
+    LEFT_SHIFT = auto()
+    LEFT_CONTROL = auto()
+    LEFT_ALT = auto()
+    LEFT_SUPER = auto()
+    LEFT_HYPER = auto()
+    LEFT_META = auto()
+    RIGHT_SHIFT = auto()
+    RIGHT_CONTROL = auto()
+    RIGHT_ALT = auto()
+    RIGHT_SUPER = auto()
+    RIGHT_HYPER = auto()
+    RIGHT_META = auto()
+    ISO_LEVEL3_SHIFT = auto()
+    ISO_LEVEL5_SHIFT = auto()
+
+
+Keys = LatinKeys | FunctionalKeys
